requirements.txt
requests
python-dotenv
ffmpeg-python
guessit
pycountry